PR: SCM-72 Submitted by : David Hawkins Reviewed by : Emmanuel Venisse
Start of work. 
o moving the revision/tag/branch methods out of the SvnUpdateCommand and
created a SvnTagBranchUtils (also the same for the test cases).
o ScmProviderRepository no longer populates tagBase in its constructor.  The
tagBase must be explicitly set if a user needs to use it due to a
non-standard directory layout.

Big thanks to Brian

git-svn-id: https://svn.eu.apache.org/repos/asf/maven/scm/trunk@332652 13f79535-47bb-0310-9956-ffa450edef68