PR: SCM-72 Submitted by : David Hawkins Reviewed by : Emmanuel Venisse
On the SvnChangeLogCommand command, I changed the date format specified in the command line.
The original format looked like: svn log -v -r "{2003/09/10 GMT}:{2003/10/10 GMT}"
But this format was not legal in svn version 1.2.3 on cygwin, linux, or windows. I am not sure if a previous version supported this date format, but the current does not.
The new date format looks like: svn log -v -r "{2003-09-10 00:00:00 +0000}:{2003-10-10 00:00:00 +0000}"
It is now specifying the time along with the date. The api allows a user to pass a date, so we should probably use the time in the date as well.
I also changed ScmTestCase.getDate() to support passing time values so I could test the changelog command creation.


git-svn-id: https://svn.apache.org/repos/asf/maven/scm/trunk@344103 13f79535-47bb-0310-9956-ffa450edef68