Fixed timing issue.
While parsing the svn commandline output, checks
are performed if files are files or directories.
At the time of the check, those files may not have
been written yet.

Moved generic code to a new abstract base class,
and added a method that filters the file list
after the execution of the command is complete.


git-svn-id: https://svn.apache.org/repos/asf/maven/scm/trunk@453271 13f79535-47bb-0310-9956-ffa450edef68