[SCM-488] fail to release if not on a remote HEAD This was mainly caused by a wrong git-pull instead of using git-checkout. Very similar to SCM-502
[SCM-498] instead of simply using "origin" we will now use fetch and push URLs as specified in the connection URL
* The sanity checks have been removed because they are way too tight.
E.g. it has not been possible to use an URL like httpuser@server.com
because the default protocol is git in this case and not http.
   
* The GitScmProviderRepository#getUrl() has been replaced with 2 functions
getFetchUrl() and getPushURL()
    
* The whole URL parsing algorithm has been fixed and should now be in sync
with the URL format documented in git-fetch.
    
* URL reconstruction now happens at runtime. So any setUser and setPassword
after the creation of the ScmProvider will get used.
    
* a few test cases have been fixed. Additional tests for fetch/push have
been added.
    
* No usage of "origin" as remote shortcut anymore. They all have been
replaced with either getFetchUrl() or getPushUrl() accordingly.

* We now also specify the branch when pushing to the upstream repository or fetching from it.

[SCM-502]  when checking out a tag we must do a git-fetch instead of a git-pull
* This has do be done since the tag may be on another branch as the current
branch. Performing a git-pull failed in some cases witha 'cannot merge'
exception and if not it may have created an invalid merge in your local repository.
Also a tag in git may be on multiple branches. This happens if the branching
happens after the tag got created.
    
* Checking out a tag will now create a 'detached HEAD' as specified by the
git documentation.



git-svn-id: https://svn.apache.org/repos/asf/maven/scm/trunk@823147 13f79535-47bb-0310-9956-ffa450edef68