[SCM-503] create a native Java GIT provider using JGit
Please note that we still have to wait for jgit-core to be released!
In the meantime, I commented out the jgit provider from the <modules>.
After eclipse released jgit-core, we can move on and release jgit-simple
(hosted on sonatype due to it's BSD license nature) and finally enable
this jgit provider for maven.


git-svn-id: https://svn.apache.org/repos/asf/maven/scm/trunk@894145 13f79535-47bb-0310-9956-ffa450edef68