SCM-797: Workaround for Windows command line length limitation
As suggested by PLXUTILS-107 and implemented by gwt-maven-plugin, it is
possible to avoid the Windows command line length limitation by invoking
the executable directly instead of via a shell. This commit adapts the
gwt-maven-plugin implementation to supply a fake Shell in Windows
environments in order to invoke the Git executable directly.